Add support for rendering from local file and inline template
This commit adds support for pointing to a local ftl-file on your
system instead of in the deployed application. This is only intended
to be used in local development of ftl-files. By pointing to the
local file, instead of the deployed file the developer needs only
refresh the page to see changes, instead of having to redeploy the
entire application.

It also adds support for using a string (as opposed to a file) as a
template.